+++ CHANGELOG.rst
@@ -13,7 +13,8 @@ Change Log
 * Adds endpoints for target search by carton and program
 * Adds ``db_reset`` option to avoid closing and resetting the database connection
 * Updates main, cone, and carton/program search to use new ``has_been_observed`` column; default is True.
-* Updates the ``append_pipes`` query to return the new ``has_been_observed`` column
+* Updates the ``append_pipes`` query to return the new ``has_been_observed`` column.
+* Include parent catalog associations in the ``/target/catalogs`` endpoint (#37).
 
 0.1.0 (10-24-2023)
 ------------------